Moved message.ts to type-tests.ts
